% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_lda_r2}
\alias{calc_lda_r2}
\title{Calculate R-squared for a tidylda Model}
\usage{
calc_lda_r2(dtm, theta, beta, threads)
}
\arguments{
\item{dtm}{must be of class dgCMatrix}

\item{theta}{a theta matrix}

\item{beta}{a beta matrix}

\item{threads}{number of parallel threads}
}
\value{
Numeric scalar between negative infinity and 1
}
\description{
Formats inputs and hands off to \link[mvrsquared]{calc_rsquared}
}
\keyword{internal}
