% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_dtm}
\alias{convert_dtm}
\title{Convert various things to a \code{dgCMatrix} to work with various functions 
and methods}
\usage{
convert_dtm(dtm)
}
\arguments{
\item{dtm}{the data you want to convert}
}
\value{
an object of class \code{dgCMatrix}
}
\description{
Presently, \code{tidylda} makes heavy usage of the \code{dgCMatrix} class.
  However, a user may have created a DTM (or TCM) in one of several classes.
  Since data could be in several formats, this function converts them to a 
  \code{dgCMatrix} before passing them along.
}
\keyword{internal}
