% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_lda_c}
\alias{fit_lda_c}
\title{Main C++ Gibbs sampler for Latent Dirichlet Allocation}
\usage{
fit_lda_c(
  docs,
  Nk,
  beta,
  alpha,
  Cd,
  Cv,
  Ck,
  Zd,
  Phi,
  iterations,
  burnin,
  freeze_topics,
  calc_likelihood,
  optimize_alpha
)
}
\arguments{
\item{docs}{List with one element for each document and one entry for each token
as formatted by \code{\link[tidylda]{initialize_topic_counts}}}

\item{Nk}{int number of topics}

\item{beta}{NumericMatrix for prior of tokens over topics}

\item{alpha}{NumericVector prior for topics over documents}

\item{Cd}{IntegerMatrix denoting counts of topics in documents}

\item{Cv}{IntegerMatrix denoting counts of tokens in topics}

\item{Ck}{IntegerVector denoting counts of topics across all tokens}

\item{Zd}{List with one element for each document and one entry for each token
as formatted by \code{\link[tidylda]{initialize_topic_counts}}}

\item{Phi}{NumericMatrix denoting probability of tokens in topics}

\item{iterations}{int number of gibbs iterations to run in total}

\item{burnin}{int number of burn in iterations}

\item{freeze_topics}{bool if making predictions, set to \code{TRUE}}

\item{calc_likelihood}{bool do you want to calculate the log likelihood each iteration?}

\item{optimize_alpha}{bool do you want to optimize alpha each iteration?}
}
\description{
This is the C++ Gibbs sampler for LDA. "Abandon all hope, ye who enter here."
}
\keyword{internal}
