% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_beta}
\alias{format_beta}
\title{Format \code{beta} For Input into \code{fit_lda_c}}
\usage{
format_beta(beta, k, Nv)
}
\arguments{
\item{beta}{the prior for words over topics. Can be a numeric scalar, numeric
vector, or numeric matrix.}

\item{k}{the number of topics.}

\item{Nv}{the total size of the vocabulary as inherited from \code{ncol(dtm)}
in \code{\link[tidylda]{tidylda}}.}
}
\value{
Returns a list with two elements: \code{beta} and \code{beta_class}.
  \code{beta} is the post-formatted version of \code{beta} in the form of a
  \code{k} by \code{Nv} numeric matrix. \code{beta_class} is a character
  denoting whether or not the user-supplied \code{beta} was a "scalar",
  "vector", or "matrix".
}
\description{
There are a bunch of ways users could format \code{beta} but the C++ Gibbs
  sampler in \code{\link[tidylda]{fit_lda_c}} only takes it one way. This function does the
  appropriate formatting. It also returns errors if the user input a malformatted
  \code{beta}.
}
\keyword{internal}
