% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.tidylda.R
\name{generate_sample}
\alias{generate_sample}
\title{Generate a sample of LDA posteriors}
\usage{
generate_sample(dir_par, matrix, times)
}
\arguments{
\item{dir_par}{matrix of Dirichlet hyperparameters, one column per}

\item{matrix}{character of "theta" or "beta", indicating which posterior
matrix \code{dir_par}'s columns are from.}

\item{times}{Integer, number of samples to draw.}
}
\value{
Returns a tibble with one row per parameter per sample.
}
\description{
Helper function called by both posterior.tidylda and predict.tidylda to
  generate samples from the posterior.
}
\keyword{internal}
