% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tidylda_model.R
\name{print.tidylda_model}
\alias{print.tidylda_model}
\title{Print Method for tidylda_model}
\usage{
\method{print}{tidylda_model}(x, digits = max(3L, getOption("digits") -
  3L), ...)
}
\arguments{
\item{x}{an object of class \code{tidylda_model}}

\item{digits}{minimal numer of significant digits}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print a summary for objects of class \code{tidylda_model}
}
\examples{
\dontrun{
dtm <- textmineR::nih_sample_dtm

lda <- fit_tidylda(dtm = dtm, k = 10, iterations = 100)

print(lda)

lda

print(lda, digits = 2)
}
}
