% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidylda-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{augment}
\alias{glance}
\alias{tidy}
\alias{refit}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics:augment]{augment()}}, \code{\link[generics:glance]{glance()}}, \code{\link[generics:refit]{refit()}}, \code{\link[generics:tidy]{tidy()}}}
}}

