% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy.dgCMatrix}
\alias{tidy.dgCMatrix}
\title{Create a tidy tibble for a dgCMatrix}
\usage{
\method{tidy}{dgCMatrix}(x, ...)
}
\arguments{
\item{x}{must be of class dgCMatrix}

\item{...}{Extra arguments, not used}
}
\description{
Create a tidy tibble for a dgCMatrix. Will probably be a PR to
  \link[tidytext]{tidytext} in the future
}
\keyword{internal}
