% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-methods.R
\name{tidy.tidylda}
\alias{tidy.tidylda}
\title{Tidy a matrix from a \code{tidylda} topic model}
\usage{
\method{tidy}{tidylda}(x, matrix, log = FALSE, ...)
}
\value{

}
\description{
Tidy a matrix from a \code{tidylda} topic model
}
