% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_dgcmatrix}
\alias{tidy_dgcmatrix}
\title{Create a tidy tibble for a dgCMatrix}
\usage{
tidy_dgcmatrix(x, ...)
}
\arguments{
\item{x}{must be of class dgCMatrix}

\item{...}{Extra arguments, not used}
}
\value{
Returns a triplet matrix with columns "document", "term", and "count"
}
\description{
Create a tidy tibble for a dgCMatrix. Will probably be a PR to
  \link[tidytext]{tidytext} in the future
}
\keyword{internal}
