% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_lda_r2}
\alias{calc_lda_r2}
\title{Calculate R-squared for a tidylda Model}
\usage{
calc_lda_r2(dtm, theta, phi, batch_size, ...)
}
\arguments{
\item{dtm}{must be of class dgCMatrix}

\item{theta}{a theta matrix}

\item{phi}{a phi matrix}

\item{batch_size}{for parallel processing}

\item{...}{other aruments passed to \code{\link[furrr]{future_map}}}
}
\description{
Formats inputs and hands off to mvrsquared::calc_rsquared
}
\keyword{internal}
