% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-methods.R
\name{glance.tidylda}
\alias{glance.tidylda}
\title{Glance method for \code{tidylda} objects}
\usage{
\method{glance}{tidylda}(x, ...)
}
\arguments{
\item{x}{an object of class \code{tidylda}}

\item{...}{other arguments passed to methods}
}
\description{
\code{glance} constructs a single-row summary "glance" of a \code{tidylda}
  topic model.
}
\examples{
\donttest{
  dtm <- textmineR::nih_sample_dtm
  
  lda <- tidylda(dtm = dtm, k = 10, iterations = 100, burnin = 75)
  
  glance(lda)
  
}
}
