% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_tidylda.R
\name{new_tidylda}
\alias{new_tidylda}
\title{Construct a new object of class \code{tidylda}}
\usage{
new_tidylda(phi, theta, gamma, alpha, beta, summary, call, ...)
}
\arguments{
\item{phi}{a numeric matrix whose rows are the posterior estimates
of P(token|topic)}

\item{theta}{a numeric matrix  whose rows are the posterior estimates of
P(topic|document)}

\item{gamma}{a numeric matrix whose rows are the posterior estimates of
P(topic|token)}

\item{alpha}{either a numeric scalar or numeric vector of length ncol(theta);
the prior for topics over documents}

\item{beta}{either a numeric scalar, numeric vector of length ncol(phi), or
numeric matrix with dimensions dim(phi); the prior for tokens over topics.}

\item{summary}{a \code{\link[tibble]{tibble}} with the following columns:
\code{topic} (numeric), \code{prevalence} (numeric), \code{coherence} (numeric),
\code{top_terms} (character)}

\item{call}{the function call used to create the model. See note, below.}

\item{...}{Other elements you would like to populate in the \code{tidylda} object.}
}
\value{
Returns an S3 object of class \code{tidylda} with the following slots:

  \code{phi} is a numeric matrix whose rows are the posterior estimates
    of P(token|topic)

  \code{theta} is a numeric matrix  whose rows are the posterior estimates of
    P(topic|document)

  \code{gamma} is a numeric matrix whose rows are the posterior estimates of
    P(topic|token). This is calculated by making a call to
    \code{link[textmineR]{CalcGamma}} which uses Bayes's rule to calculate
    \code{gamma} from \code{phi}, \code{theta}, and P(document) (which is
    proportional to \code{Matrix::rowSums(dtm)}).

  \code{alpha} is the prior for topics over documents. If \code{optimize_alpha}
    is \code{FALSE}, \code{alpha} is what the user passed when calling
    \code{\link[tidylda]{tidylda}}. If \code{optimize_alpha} is \code{TRUE},
    \code{alpha} is a numeric vector returned in the \code{alpha} slot from a
    call to \code{\link[tidylda]{fit_lda_c}}.

  \code{beta} is the prior for tokens over topics. This is what the user passed
    when calling \code{\link[tidylda]{tidylda}}.

  \code{summary} is the result of a call to \code{\link[tidylda]{summarize_topics}}

  \code{call} is the result of \code{\link[base]{match.call}} called at the top
    of \code{\link[tidylda]{tidylda}}

  \code{log_likelihood} is a \code{\link[tibble]{tibble}} whose columns are
    the iteration and log likelihood at that iteration. This slot is only populated
    if \code{calc_likelihood = TRUE}

  \code{r2} is a numeric scalar resulting from a call to
    \code{\link[textmineR]{CalcTopicModelR2}}. This slot only populated if
    \code{calc_r2 = TRUE}
}
\description{
This is a helper function for creating new \code{tidylda} objects.
}
\note{
You can technically pass a \code{\link[base]{data.frame}} as the \code{summary}
  parameter. However, you may experience unintended side effects when using
  \code{\link[tidylda]{print.tidylda}}.

  The class of \code{call} isn't checked. It's just passed through to the
  object returned by this function. Might be useful if you are using this
  function for troubleshooting or something.
}
\keyword{internal}
