% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tidylda.R
\name{print.tidylda}
\alias{print.tidylda}
\title{Print Method for tidylda}
\usage{
\method{print}{tidylda}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class \code{tidylda}}

\item{digits}{minimal numer of significant digits}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print a summary for objects of class \code{tidylda}
}
\examples{
\donttest{
dtm <- textmineR::nih_sample_dtm

lda <- tidylda(dtm = dtm, k = 10, iterations = 100)

print(lda)

lda

print(lda, digits = 2)
}
}
