% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidylda-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{augment}
\alias{glance}
\alias{tidy}
\alias{refit}
\alias{colSums}
\alias{colMeans}
\alias{rowSums}
\alias{rowMeans}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{glance}}, \code{\link[generics]{refit}}, \code{\link[generics]{tidy}}}

  \item{Matrix}{\code{\link[Matrix:colSums]{colMeans}}, \code{\link[Matrix]{colSums}}, \code{\link[Matrix:colSums]{rowMeans}}, \code{\link[Matrix:colSums]{rowSums}}}
}}

