% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidylda.R
\docType{package}
\name{tidylda-package}
\alias{tidylda-package}
\title{tidylda}
\description{
Latent Dirichlet Allocation Using 'tidyverse' Conventions
}
\details{
Implements the WarpLDA algorithm for Latent Dirichlet Allocation 
using style conventions from the 'tidyverse' and specifically 'tidymodels'. 
Designed to work seamlessly with 'textmineR'.
}
