% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidylda.R
\docType{package}
\name{tidylda}
\alias{tidylda}
\title{Latent Dirichlet Allocation Using 'tidyverse' Conventions}
\description{
Implements an algorithm for Latent Dirichlet Allocation (LDA)
  using style conventions from the 'tidyverse' and specifically 'tidymodels'. 
  Also implements several novel features for LDA such as guided models and 
  transfer learning.
}
\keyword{internal}
