% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidylda-fit-methods.R
\name{tidylda_bridge}
\alias{tidylda_bridge}
\title{Bridge function for fitting \code{tidylda} topic models}
\usage{
tidylda_bridge(
  dtm,
  k,
  iterations,
  burnin,
  alpha,
  beta,
  optimize_alpha,
  calc_likelihood,
  calc_r2,
  return_data,
  mc,
  ...
)
}
\description{
Takes in arguments from various \code{tidylda} S3 methods and fits the 
  resulting topic model. The arguments to this function are documented in
  \code{\link[tidylda]{tidylda}}.
}
\keyword{internal}
