% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.tidylda_model.R
\name{update}
\alias{update}
\title{Update methods for topic models}
\usage{
update(object, ...)
}
\arguments{
\item{object}{An existing trained topic model}

\item{...}{Additional arguments to the call}
}
\description{
\code{update} updates a previously-trained topic model based
on new data or continues training a model with its original data. Useful for 
updates or transfer learning.
}
